ENH: Add pipx variant of pip3
In Ubuntu 24.04, pip3 is more insistent that adding packages to the global Python path is wrong.
> This environment is externally managed
I've implemented several workarounds to differentiate between natively packaged and Python-installed packages, but this led me to a renewed search for alternative solutions, and this (now) exists in the form of pipx.
- https://stackoverflow.com/a/78652149/813602
- https://python.land/virtual-environments/pipx

pipx installs Python packages that include a CLI (like mypy, HTTPie, or yt-dlp) in segregated environments (since version 1.5.0 globally under /opt/pipx) and adds the launcher script to /usr/local/bin/.
That's exactly what I need.
The logic works just like with pip3; however, there are slight differences in how optional modules are specified (as part of the package, not separately) and installed (all together, not accummulatingly, so we have to merge existing with added modules ourselves).

WIP: Still missing: pipx+url type, and the ability to "pipx inject" additional dependencies