pipx: ENH: Store based on the (pure) package name
In order to add the pipx+url: type, there we only have the PACKAGE-NAME, not a full PACKAGE-SPEC (or to be exact the PACKAGE-SPEC is a downloaded file system location that doesn't correspond to what's given to pipx:).
To detect packages installed by pipx: and vice versa, we need to key on the pure package name.
pipx-list-packages now supports --both-package-name-and-spec to return both.
Use installedPipxPackageSpecToName and installedPipxPackageNameToSpec to cache and translate.
For installing, we need to use the original package spec, but inject the combined dependency modules. (I hope this works ;-)