pip3: Avoid "This environment is externally managed" error via --break-system-packages
We now have an alternative for Python CLI apps via the pipx type, but (unless at least version 1.5 lands in the package repos), the pip3 type still is needed to install it.
On Ubuntu 20.04, pip3 just warns, but on Ubuntu 24.04, this is an error that _must_ be overridden via that command-line flag.