setup-software: BUG: update-software only says that "definitions would be installed"
When a type of the to-be-installed items needs a prerequisite installed via isAvailableOrUserAccepts{Group,Native}(), but there's no answer to recall, this results in failure, but nothing is printed and there are no install commands, but we have a non-zero addedDefinitionCount.
Check for success of addItems() and print an error if that fails.