setup-software: ENH: Record past selected definitions and accept those (on --recall-only / update-software) when decision is outdated
My setups and groupings are now so sophisticated that I don't individually accept each and every definition. It's enough to accept certain groups, and then accept all definitions from all groups for the current day. In other words, the selection now purely happens on the group level, and is much faster.
This causes a problem with update-software, however. Because it only recalled the memoized decisions, those expire after one day, and they are not considered for updates (i.e. when I add a dependency or additional action, it won't be offered).
> 1 not yet installed definition was declined.
I need to accept any previously selected definition (regardless of the memoized validity period that may have been expired) in order to offer updates. (Same for previously selected definition groups, but that'll come later.)
I can't simply ignore the memoization process altogether, as that would also indiscriminately add newly added definitions.
Therefore, a different database is needed that stores all accepted definitions (by subject), and that takes precedence when recalling.
Modify the --recall-only behavior (so that update-software doesn't need to be modified), and keep the original behavior (which would only make sense when redoing a recent software selection) via the new --recall-current-only alternate flag.