setup-software: ENH: Record past selected definition groups and accept those (on --group-recall-only / update-software) when decision is outdated (cp. fcde44f)
The selection expiry problem also affects definition groups, as those may also only be accepted for a short duration (1 day). The problem is less pronounced here, as it generally has less consequences to permanently accept a group. (Maybe we just want one definition from a group, but don't want to be questioned again for anything from that group later?)
As group selection is implemented just like individual definitions (with -groups added to the memoization group), adding this is straightforward, now that we have extracted the askGroup() function.
Modify the --group-recall-only behavior (so that update-software doesn't need to be modified), and keep the original behavior (which would only make sense when redoing a recent software selection) via the new --group-recall-current-only alternate flag.
Refactoring: Move the storeSelectedDefinition() into askTo() (and likewise storeSelectedDefinitionGroup() into askGroup()) so that it's not duplicated. This also adds the forgotten coverage of askToInstall().

(This relates to commit fcde44f5ff7b75b11eda12996977e737acadcd8b)