BUG: "enable delta beautified diffs in Git" isn't enabled via dummy
Because addDummy() is executed before addDebUrl(), so the check whether the item has been selected doesn't find it.
We've lumped in the dummies together with the configurations, but the dummies actually have to be added _after_ the items, not before them.
Introduce a separate missingItemDummies list.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release